requirements.txt
# Core Django dependencies
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Background tasks
celery==5.3.4
redis==5.0.1

# LangChain and AI dependencies
langgraph>=0.5.4
langchain-community>=0.3.9
langchain-openai>=0.3.28
langchain-anthropic>=0.3.15
langchain-tavily
langsmith>=0.3.37
openai>=1.99.2
tavily-python>=0.5.0

# File processing
PyPDF2==3.0.1
python-magic==0.4.27

# Environment management
python-decouple==3.8

# Testing
pytest==7.4.3
pytest-django==4.7.0
hypothesis==6.92.1

# Development
black==23.12.1
flake8==6.1.0